Introduce unique book id for faster and easier lookup

+++ app/database/migrations/2014_04_04_131602_create_books_table.php
@@ -28,7 +28,7 @@ class CreateBooksTable extends Migration
         });
 
         Schema::create('books', function (Blueprint $table) {
-            $table->tinyInteger('id')->unsigned();
+            $table->integer('id')->unsigned();
             $table->timestamps();
             $table->unsignedInteger('translation_id')->foreign('translation_id')->references('id')->on('translations');
             $table->primary(['id', 'translation_id']);
